+++ crates/zng-view-api/src/lib.rs
@@ -404,7 +404,7 @@ declare_api! {
     /// Returns the new image texture ID. If the `image_id` is not loaded returns the [`INVALID`] texture ID.
     ///
     /// [`INVALID`]: ImageTextureId::INVALID
-    pub fn use_image(&mut self, id: WindowId, image_id: ImageId) -> ImageTextureId; // !!: TODO size hints?
+    pub fn use_image(&mut self, id: WindowId, image_id: ImageId) -> ImageTextureId;
 
     /// Replace the image resource in the window renderer.
     ///