+++ crates/zng-view/src/image_cache.rs
@@ -134,35 +134,41 @@ impl ImageCache {
         rayon::spawn(move || {
             let r = match format {
                 ImageDataFormat::Bgra8 { size, density } => {
+                    // is already decoded bgra8
+
                     let expected_len = size.width.0 as usize * size.height.0 as usize * 4;
                     if data.len() != expected_len {
                         Err(formatx!(
                             "pixels.len() is not width * height * 4, expected {expected_len}, found {}",
                             data.len()
                         ))
-                    } else if mask.is_some() {
-                        // !!: TODO this is not right? Its Bgra8
-                        let r = Self::convert_decoded(
-                            image::DynamicImage::ImageLuma8(
-                                image::ImageBuffer::from_raw(size.width.0 as _, size.height.0 as _, data.to_vec()).unwrap(),
-                            ),
-                            mask,
-                            density,
-                            None,
-                            downscale,
-                            &resizer,
-                        );
-                        match r {
-                            Ok((pixels, size, _, is_opaque, _)) => Ok((pixels, size, density, is_opaque, true)),
+                    } else if let Some(mask) = mask {
+                        // but is used as mask, convert to a8
+
+                        Self::convert_bgra8_to_mask(size, &data, mask, density, downscale, &resizer).map_err(|e| e.to_txt())
+                    } else {
+                        // and is used as bgra8, downscale if needed
+                        match Self::downscale_decoded(mask, downscale, &resizer, size, &data) {
+                            Ok(downscaled) => match downscaled {
+                                Some((size, data_mut)) => match data_mut.finish_blocking() {
+                                    Ok(data) => {
+                                        let is_opaque = data.chunks_exact(4).all(|c| c[3] == 255);
+                                        Ok((data, size, None, is_opaque, true))
+                                    }
+                                    Err(e) => Err(e.to_txt()),
+                                },
+                                None => {
+                                    let is_opaque = data.chunks_exact(4).all(|c| c[3] == 255);
+                                    Ok((data, size, None, is_opaque, true))
+                                }
+                            },
                             Err(e) => Err(e.to_txt()),
                         }
-                    } else {
-                        // !!: TODO downscale
-                        let is_opaque = data.chunks_exact(4).all(|c| c[3] == 255);
-                        Ok((data, size, density, is_opaque, false))
                     }
                 }
                 ImageDataFormat::A8 { size } => {
+                    // is already decoded mask
+
                     let expected_len = size.width.0 as usize * size.height.0 as usize;
                     if data.len() != expected_len {
                         Err(formatx!(
@@ -170,6 +176,7 @@ impl ImageCache {
                             data.len()
                         ))
                     } else if mask.is_none() {
+                        // but is used as mask, convert to bgra8
                         let r = Self::convert_decoded(
                             image::DynamicImage::ImageLuma8(
                                 image::ImageBuffer::from_raw(size.width.0 as _, size.height.0 as _, data.to_vec()).unwrap(),
@@ -185,17 +192,35 @@ impl ImageCache {
                             Err(e) => Err(e.to_txt()),
                         }
                     } else {
-                        let is_opaque = data.iter().all(|&c| c == 255);
-                        Ok((data, size, None, is_opaque, true))
+                        // and is used as mask, downscale if needed
+                        match Self::downscale_decoded(mask, downscale, &resizer, size, &data) {
+                            Ok(downscaled) => match downscaled {
+                                Some((size, data_mut)) => match data_mut.finish_blocking() {
+                                    Ok(data) => {
+                                        let is_opaque = data.iter().all(|&c| c == 255);
+                                        Ok((data, size, None, is_opaque, true))
+                                    }
+                                    Err(e) => Err(e.to_txt()),
+                                },
+                                None => {
+                                    let is_opaque = data.iter().all(|&c| c == 255);
+                                    Ok((data, size, None, is_opaque, true))
+                                }
+                            },
+                            Err(e) => Err(e.to_txt()),
+                        }
                     }
                 }
                 fmt => {
+                    // needs decoding
+
                     #[cfg(not(feature = "image_any"))]
                     {
                         let _ = (max_decoded_len, downscale);
                         Err(zng_txt::formatx!("no decoder for {fmt:?}"))
                     }
 
+                    // identify codec and parse header metadata
                     #[cfg(feature = "image_any")]
                     match Self::header_decode(&fmt, &data[..]) {
                         Ok(h) => {
@@ -206,6 +231,7 @@ impl ImageCache {
                                     "image {size:?} needs to allocate {decoded_len} bytes, but max allowed size is {max_decoded_len} bytes",
                                 ))
                             } else {
+                                // notify metadata already
                                 if let Some(d) = downscale {
                                     size = d.resize_dimensions(size);
                                 }
@@ -215,7 +241,10 @@ impl ImageCache {
                                     density: h.density,
                                     is_mask: false,
                                 }));
+
+                                // decode
                                 match Self::image_decode(&data[..], h.format, downscale, h.orientation) {
+                                    // convert to bgra8 and downscale
                                     Ok(img) => match Self::convert_decoded(img, mask, h.density, h.icc_profile, downscale, &resizer) {
                                         Ok(r) => Ok(r),
                                         Err(e) => Err(e.to_txt()),